Refactor: Move egexValidator from desktop to core
